Create CustomEntityFirework.java
Spawn Exploding Fireworks!

Finally, the code we've all been waiting for. An instant firework explosion effect!
Finally, a reliable way to launch exploding fireworks.
No launch sound
No visible rocket (well, for 1 tick, but the firework effect basically hides it)
Just the explosion!
Works 100% of the time!
You can send the firework effect to specific people, and the other people in the world won't see it (but those other people will, for about 1 tick, see the rocket and a trail).